Module zune_inflate::constants

Constants

DEFLATE_BLOCKTYPE_DYNAMIC_HUFFMAN

DEFLATE_BLOCKTYPE_RESERVED

DEFLATE_BLOCKTYPE_STATIC

DEFLATE_BLOCKTYPE_UNCOMPRESSED

DEFLATE_MAX_CODEWORD_LENGTH

Maximum codeword length across all codes.

DEFLATE_MAX_LITLEN_CODEWORD_LENGTH

DEFLATE_MAX_MATCH_LEN

Maximum length of a deflate match

DEFLATE_MAX_NUM_SYMS

Maximum number of symbols across all codes

DEFLATE_MAX_OFFSET_CODEWORD_LENGTH

DEFLATE_MAX_PRE_CODEWORD_LEN

Maximum codeword length in bits for each precode

DEFLATE_NUM_LITLEN_SYMS

DEFLATE_NUM_OFFSET_SYMS

DEFLATE_NUM_PRECODE_SYMS

Number of symbols in each Huffman code. Note: for the literal/length and offset codes, these are actually the maximum values; a given block might use fewer symbols.

DELFATE_MAX_LENS_OVERRUN

Maximum possible overrun when decoding codeword lengths

FASTCOPY_BYTES

Number of bytes copied per every loop

FASTLOOP_MAX_BYTES_WRITTEN

Worst case maximum number of output bytes writtern during each iteration of the fastloop.

HUFFDEC_END_OF_BLOCK

End of block entry in litlen decode table

HUFFDEC_EXCEPTIONAL

Presence of HUFFDEC_SUITABLE_POINTER or HUFFDEC_END_OF_BLOCK

HUFFDEC_LITERAL

Presence of a literal entry

HUFFDEC_SUITABLE_POINTER

Pointer entry in the litlen or offset decode table

LITLEN_DECODE_BITS

LITLEN_ENOUGH

LITLEN_TABLE_BITS

OFFSET_ENOUGH

Note, default libdeflate value is 402, but with 512, we can remove a branch check by simply doing & 511, and I’ll take that.

OFFSET_TABLEBITS

Maximum bits found in the lookup table for offsets offsets larger than this require a lookup into a sub-table

PRECODE_ENOUGH

PRECODE_TABLE_BITS

Statics

DEFLATE_PRECODE_LENS_PERMUTATION

Order which precode lengths are stored

LITLEN_DECODE_RESULTS

OFFSET_DECODE_RESULTS

PRECODE_DECODE_RESULTS

Format for precode decode table entries, Bits not explicitly contain zeroes

Functions

construct_litlen_decode_table 🔒

entry 🔒

make_precode_static_table 🔒