Module zune_inflate::decoder

Structs

DeflateDecoder

A deflate decoder instance.

DeflateHeaderTables 🔒

DeflateOptions

Options that can influence decompression in Deflate/Zlib/Gzip

Constants

RESIZE_BY 🔒

Functions

resize_and_push 🔒

Resize vector if its current space wont be able to store a new byte and then push an element to that new space