Struct zune_inflate::DeflateOptions

pub struct DeflateOptions {
    limit: usize,
    confirm_checksum: bool,
    size_hint: usize,
}

Options that can influence decompression in Deflate/Zlib/Gzip

To use them, pass a customized options to the deflate decoder.

Fields
limit: usizeconfirm_checksum: boolsize_hint: usize
Implementations

impl DeflateOptions

pub const fn get_limit(&self) -> usize

Get deflate/zlib limit option

The decoder won’t extend the inbuilt limit and will return an error if the limit is exceeded

Returns

The currently set limit of the instance

Note

This is provided as a best effort, correctly quiting is detrimental to speed and hence this should not be relied too much.

pub fn set_limit(self, limit: usize) -> Self

Set a limit to the internal vector used to store decoded zlib/deflate output.

Arguments

limit: The new decompressor limit

Returns

A modified version of DeflateDecoder

Note

This is provided as a best effort, correctly quiting is detrimental to speed and hence this should not be relied too much

pub const fn get_confirm_checksum(&self) -> bool

Get whether the decoder will confirm a checksum after decoding

pub fn set_confirm_checksum(self, yes: bool) -> Self

Set whether the decoder should confirm a checksum after decoding

Note, you should definitely confirm your checksum, use this with caution, otherwise data returned may be corrupt

Arguments

yes: When true, the decoder will confirm checksum when false, the decoder will skip checksum verification

Notes

This does not have an influence for deflate decoding as it does not have a checksum

pub const fn get_size_hint(&self) -> usize

Get the default set size hint for the decompressor

The decompressor initializes the internal storage for decompressed bytes with this size and will reallocate the vec if the decompressed size becomes bigger than this, but when the user currently knows how big the output will be, can be used to prevent unnecessary re-allocations

pub const fn set_size_hint(self, hint: usize) -> Self

Set the size hint for the decompressor

This can be used to prevent multiple re-allocations